@@ -11,7 +11,7 @@ namespace ConcertBuddy.ConsoleApp.TelegramBot.Command
     {
         private ILogger<UsageCommand> _logger = ServiceProviderSingleton.Source.GetService<ILogger<UsageCommand>>();
 
-        public UsageCommand(ISearchHandler searchHandler, ITelegramBotClient telegramBotClient, Message data) 
+        public UsageCommand(ISearchHandler searchHandler, ITelegramBotClient telegramBotClient, Message data)
             : base(searchHandler, telegramBotClient, data)
         {
         }
