@@ -4,7 +4,7 @@ using SetlistFmAPI.Models;
 
 namespace SetlistFmAPI
 {
-    public class SetlistFmClient: ISetlistFmClient
+    public class SetlistFmClient : ISetlistFmClient
     {
 
         private readonly ILogger<SetlistFmClient> _logger;
@@ -28,7 +28,7 @@ namespace SetlistFmAPI
         /// <inheritdoc />
         public async Task<Artists> SearchArtists(Artist searchFields, int page = 1)
         {
-            return await _httpClient.Load<Artists>(SetlistFmUrls.Artists(searchFields, page:page), _apiKey, _language);
+            return await _httpClient.Load<Artists>(SetlistFmUrls.Artists(searchFields, page: page), _apiKey, _language);
         }
 
         /// <inheritdoc />
